test-requirements.txt
# Testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-html==4.1.1

# Test utilities
responses==0.23.3
requests-mock==1.11.0
factory-boy==3.3.0
faker==20.1.0

# Performance testing
locust==2.17.0
memory-profiler==0.61.0
psutil==5.9.6

# Docker testing
docker==6.1.3
testcontainers==3.7.1

# Flask testing
Flask-Testing==0.8.1

# Database testing
pytest-postgresql==5.0.0

# Selenium for E2E testing
selenium==4.15.2
webdriver-manager==4.0.1

# Additional utilities
freezegun==1.2.2
parameterized==0.9.0